test_requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
attrs==21.2.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
crypto==1.4.1
cryptography==35.0.0
cycler==0.10.0
Deprecated==1.2.13
Django==3.2.9
firebase==3.0.1
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
gitdb==4.0.9
GitPython>=3.1.18
google-api-core==2.2.2
google-auth==2.0.0
google-cloud-core==2.1.0
google-cloud-datastore==2.3.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-re2==0.2.20210901
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jwcrypto==1.0
kiwisolver>=1.3.0
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
Naked==0.1.31
npm==0.1.1
numpy>=1.19.1
oauth2client==4.1.3
optional-django==0.1.0
packaging==21.2
Pillow==8.3.2
pluggy==1.0.0
proto-plus==1.19.7
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
PyJWT>2.0.0,<=3.0.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-jwt==3.3.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
semver==2.13.0
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
sseclient==0.0.27
toml==0.10.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
validators==0.18.2
Werkzeug==2.0.2
wrapt==1.13.3
